[CICO-5] separated AuthObject Class, Documented the functions and cleaned the code and outputs
